Add files via upload
This code is in the .ipynb format because it allows for you to run only certain parts of the code. I will also try to upload a .py file but it might not work as well. All code should have comments, but let me know if you have questions on how to use it.